lsp: combine file seeks and only take lines we need
